Extracted the config file loading logic into a trait.
